Profile -> Member
Includes a manually created migration.RenameModel, as the autodetector
instead directed the migration to delete the old one and create a new
model.

"Personal" views and templates where the user is editing and viewing
their own data are now prefixed by "my-member-", and the URLs for these
are now "/member/me/"

Non-personal member views and templates are prefixed with "member-"
and their URLs start with "/member/". Thus, to see your own profile
as others see it, you'd visit /member/your-username-here.