Attempt to add informed consent quiz
This isn't working for me.

Modules needed by the quiz's javascript (L7 to L10 in
public-data-enroll-3-quiz.js) cause errors when gulp runs
browserify. Commenting out L7 to L10 lets it run without error,
but when loading the page the console reports an error when
trying to use lodash: "Uncaught ReferenceError: _ is not defined"