🐛 Trigger task after upload via view logic, not signal.
The presave signal for triggering a task was broken because the S3 URL
was incorrect. (My current interpretation of this is that the file
isn't loaded into S3 until the model is saved.)

Triggering tasks via signal was originally done to trigger tasks for
data-push OAuth2 interactions. In other words, this signal was a hack
to trigger data processing after an app used the API to push user data
to us (for PGP, American Gut, etc.).

But a save-associated signal can accidentally be triggered by other
manipulations of that object. Ideally the task is triggered by view
logic specific to the process, which is what I'm implementing here.