🐛 Fix badge info for project auth
Previously this only worked if the project was an approved activity.

This commit refactors the code that compiles activity information from
a DataRequestProject so it can be called for any project (e.g. an
unapproved one). The project authorization views then use this to
retrieve and pass the appropriate 'badge' data to the authorization
templates.