Switching to global JS imports to fix issues
jQuery bundled with browserify seems to be leading to various
issues with how it's imported: causing an overwrite of a global
variable (window.jQuery) and bugs with that. Solutions that aren't
global jQuery seem to be inconsistently working, maybe due to
some race condition in loading...

Fixes #686 (again).

Parsley and Select2 both seem to depend on a require of jQuery to
be bundled and work with browserify. (See for example:
https://github.com/guillaumepotier/Parsley.js/issues/1104)
So moving jQuery to be global seems to end up needing these to also
be global.

With all these global, everything _seems_ to be working...