New redirect storage JS, simplify activity join button
For elements where we want to store the current URL to session, to return to
upon login/signup: switch to using a class and add an event handler to it
with JQuery.

This commit also greatly simplifies the activity join button to remove pop-ups
that arise when a user is not logged in, or doesn't have a verified email.

For the former case, we expect the user to hit a PrivateMixin that prompts login
during the authorization flow.