Decrease page rendering speed by reducing database queries and adds separate Add Data and Explore & share fields (#946)
Prior to this change, we were hitting the database more times than were necessary, but also we were iterating over the database results multiple time. This caused data display to take quadratic time.  This issue is documented in #942 

This changes to using single database queries and then only iterating over the results once to actually serialize the data for display.

This applies to the following endpoints:
/
/create/
/members/
/member/me/joined/

As an example of the difference this change caPrior to this change, we were hitting the database more times than were necessary, but also we were iterating over the database results multiple time. This caused data display to take quadratic time.

This changes to using single database queries and then only iterating over the results once to actually serialize the data for display.

This applies to the following endpoints:
/
/create/
/members/
/member/me/joined/

As an example of the difference this change causes, django-silk reported a load time of 1314ms overall and 515ms for the 78 db queries in order to render the home page. With this change, that becomes 683ms overall with 103 ms spent on the 14 db queries. This difference is consistent and reproducible.uses, django-silk reported a load time of 1314ms overall and 515ms for the 78 db queries in order to render the home page. With this change, that becomes 683ms overall with 103 ms spent on the 14 db queries. This difference is consistent and reproducible.

In addition, this change adds fields to manually select if a project should show up under Add data or Explore and share or both, as making that a separate change would have involved more work.

Signed-off-by: Mairi Dulaney <jdulaney@fedoraproject.org>