Overhaul of public data APIs
This creates a new set of publicly available APIs that
present public information/data about the following:
- DataFiles
- DataTypes
- Members
- Projects

The naming schemes and formats are generally re-used, and
nesting has been avoided -- for example, project serialization
doesn't list public files for each project, that's available at
another endpoint. (Nesting has caused previous endpoints to crash.)

Documentation is updated as well.

Some previous endpoints are retained, as they are known to be
in use, or might be. These are deprecated.