Add waffle, switch logging to feature flag control
This adds django-waffle, using a custom Flag model to enable managing
flags according to a "subject" (a relevant User that isn't necessarily
the request User).

The flag is then used to manage the file access logging feature,
which is incomplete and leads to heavy database use when active for
all users.