Add optional signature for webhooks
This update:

* adds an optional webhook_secret field an OAuth2 project can provide
* if provided, this is used to create an hmac-sha1 hexdigest to verify
a request made to a webhook
* this is included as a custom header "X-OpenHumans-Webhooks-Signature"
when calling a webhook
* documentation is added for the deauthorization webhook and webhook secret
* the JSON payload for the deauthorization webhook has been corrected
(it was accidentally doing a second JSON encoding)

The last item will break anything that was using this webhook. However,
the webhook wasn't actually documented before now. Support for it was added
to the django-open-humans library, and so a concurrent update to that
package is staged.