Add optional signature for webhooks (#1115)
This update:

* adds an optional webhook_secret field an OAuth2 project can provide
* if provided, this is used to create an hmac-sha1 hexdigest to verify
a request made to a webhook
* this is included as a custom header "X-OpenHumans-Webhooks-Signature"
when calling a webhook
* documentation is added for the deauthorization webhook and webhook secret
* the JSON payload for the deauthorization webhook has been corrected
(it was accidentally doing a second JSON encoding)

The webhook wasn't actually documented before now, but the last item would break something already using the webhook. So as a fallback, the server tries again with the old double-encoded version if a non-200 response is received. An update supporting the new behavior is also staged for the django-open-humans package.